+++ async-race/src/modules/components/index.ts
@@ -4,3 +4,8 @@ export { createColorChoice } from './colorInput';
 export { createPageTitle } from './pageTitle';
 export { createPageSubtitle } from './pageSubtitle';
 export { createCarImage } from './car';
+export { createFooter } from './footer';
+export { createHeader } from './header';
+export { createNavigation } from './navigation';
+export { createInputBlock } from './inputsBlock';
+export { createRaceHeaderBtns } from './raceHeaderBtns';
