+++ async-race/src/modules/features/index.ts
@@ -1,11 +1,4 @@
-import { createRaceHeaderBtns } from './raceHeaderBtns';
-
-export { createInputBlock } from './inputsBlock';
-export { createNavigation } from './navigation';
-export { createRaceHeaderBtns } from './raceHeaderBtns';
-export { createHeader } from './header';
-export { createFooter } from './footer';
 export { garageMainCreate } from './garageMain';
 export { createCarItem } from './carItem';
 export { winnersMainCreate } from './winnersMain';
-export { createTable } from './table';
+export { createTable } from './winnersTable';
